split out pyomo control models
