--- h2integrate/control/control_strategies/controller_baseclass.py
@@ -1,52 +1,0 @@
-import openmdao.api as om
-
-
-class ControllerBaseClass(om.ExplicitComponent):
-    """
-    Base class for open-loop controllers in the H2Integrate system.
-
-    This class provides a template for implementing open-loop controllers. It defines the
-    basic structure for inputs and outputs and requires subclasses to implement the `compute`
-    method for specific control logic.
-
-    Attributes:
-        plant_config (dict): Configuration dictionary for the overall plant.
-        tech_config (dict): Configuration dictionary for the specific technology being controlled.
-    """
-
-    def initialize(self):
-        """
-        Declare options for the component. See "Attributes" section in class doc strings for
-        details.
-        """
-
-        self.options.declare("driver_config", types=dict)
-        self.options.declare("plant_config", types=dict)
-        self.options.declare("tech_config", types=dict)
-
-    def setup(self):
-        """
-        Define inputs and outputs for the component.
-
-        This method must be implemented in subclasses to define the specific control inputs
-        and outputs.
-
-        Raises:
-            NotImplementedError: If the method is not implemented in a subclass.
-        """
-        raise NotImplementedError("This method should be implemented in a subclass.")
-
-    def compute(self, inputs, outputs):
-        """
-        Perform computations for the component.
-
-        This method must be implemented in subclasses to define the specific control logic.
-
-        Args:
-            inputs (dict): Dictionary of input values.
-            outputs (dict): Dictionary of output values.
-
-        Raises:
-            NotImplementedError: If the method is not implemented in a subclass.
-        """
-        raise NotImplementedError("This method should be implemented in a subclass.")
