+++ h2integrate/core/supported_models.py
@@ -97,10 +97,6 @@ from h2integrate.converters.natural_gas.natural_gas_cc_ct import (
 )
 from h2integrate.converters.hydrogen.singlitico_cost_model import SingliticoCostModel
 from h2integrate.converters.co2.marine.direct_ocean_capture import DOCCostModel, DOCPerformanceModel
-from h2integrate.control.control_strategies.pyomo_controllers import (
-    OptimizedDispatchController,
-    HeuristicLoadFollowingController,
-)
 from h2integrate.converters.hydrogen.geologic.mathur_modified import GeoH2SubsurfaceCostModel
 from h2integrate.resource.solar.nrel_developer_goes_api_models import (
     GOESTMYSolarAPI,
@@ -130,6 +126,12 @@ from h2integrate.converters.co2.marine.ocean_alkalinity_enhancement import (
 )
 from h2integrate.converters.hydrogen.custom_electrolyzer_cost_model import (
     CustomElectrolyzerCostModel,
+)
+from h2integrate.control.control_strategies.heuristic_pyomo_controller import (
+    HeuristicLoadFollowingController,
+)
+from h2integrate.control.control_strategies.optimized_pyomo_controller import (
+    OptimizedDispatchController,
 )
 from h2integrate.converters.hydrogen.geologic.aspen_surface_processing import (
     AspenGeoH2SurfaceCostModel,
